+++ front-end/src/app/safety-items/safety-items.component.ts
@@ -1,27 +1,46 @@
-import { Component } from '@angular/core';
-import { SafetyItemService } from '../services/safetyItem/safety-item.service';
+import { Component } from "@angular/core";
+import { SafetyItemService } from "../services/safetyItem/safety-item.service";
+import Swal from "sweetalert2";
 
 @Component({
-  selector: 'app-safety-items',
-  templateUrl: './safety-items.component.html',
-  styleUrls: ['./safety-items.component.scss']
+  selector: "app-safety-items",
+  templateUrl: "./safety-items.component.html",
+  styleUrls: ["./safety-items.component.scss"],
 })
 export class SafetyItemsComponent {
+  safetyItems: any[] = [];
+  dtOptions: DataTables.Settings = {};
+
   constructor(private safetyItemService: SafetyItemService) {}
-  safetyItems :any[]=[];
 
   ngOnInit(): void {
+    this.dtOptions = {
+      pagingType: "full_numbers",
+    };
     this.safetyItemsList();
   }
 
-  safetyItemsList(){
-    this.safetyItemService.getSafetyItems().subscribe((res:any)=>{ this.safetyItems=res})
+  safetyItemsList() {
+    this.safetyItemService.getSafetyItems().subscribe((res: any) => {
+      this.safetyItems = res;
+    });
   }
 
-  suppSafetyItem(id:number){
-    this.safetyItemService.deleteSafetyItem(id).subscribe((res:any)=>{
-      console.log('safety item supprimé !');
-      this.ngOnInit();
+  suppSafetyItem(id: number) {
+    Swal.fire({
+      title: "Êtes-vous sûre?",
+      showCancelButton: true,
+      confirmButtonColor: "btn btn-primary",
+      cancelButtonColor: "btn btn-secondary",
+      confirmButtonText: "Oui, supprimer!",
+      cancelButtonText: "Annuler",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.safetyItemService.deleteSafetyItem(id).subscribe((res: any) => {
+          console.log("safety item supprimé !");
+          this.ngOnInit();
+        });
+      }
     });
   }
 }
